requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.3.0
colorama==0.4.6
fastapi==0.117.1
h11==0.16.0
idna==3.10
mysql-connector-python==9.4.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.36.0
